refactor pre.update.password.user.action
