+++ components/user-mgt/org.wso2.carbon.identity.user.pre.update.password.action/pom.xml
@@ -94,14 +94,11 @@
                         </Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Private-Package>
-                            org.wso2.carbon.identity.user.pre.update.password.action.core.constant,
-                            org.wso2.carbon.identity.user.pre.update.password.action.core.execution,
-                            org.wso2.carbon.identity.user.pre.update.password.action.core.management,
-                            org.wso2.carbon.identity.user.pre.update.password.action.core.rule,
-                            org.wso2.carbon.identity.user.pre.update.password.action.internal
+                            org.wso2.carbon.identity.user.pre.update.password.action.internal.*
                         </Private-Package>
                         <Export-Package>
-                            org.wso2.carbon.identity.user.pre.update.password.action.service.model;
+                            !org.wso2.carbon.identity.user.pre.update.password.action.internal.*,
+                            org.wso2.carbon.identity.user.pre.update.password.action.api.*;
                             version="${carbon.identity.package.export.version}"
                         </Export-Package>
                         <Import-Package>
@@ -130,8 +127,8 @@
                             org.wso2.carbon.identity.rule.evaluation.model; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.identity.rule.evaluation.exception; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.identity.rule.evaluation.provider; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.user.action.service; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.user.action.service.model; version="${carbon.identity.package.import.version.range}"
+                            org.wso2.carbon.identity.user.action.api.model; version="${carbon.identity.package.import.version.range}",
+                            org.wso2.carbon.identity.user.action.api.service; version="${carbon.identity.package.import.version.range}"
                         </Import-Package>
                     </instructions>
                 </configuration>
