+++ components/user-mgt/org.wso2.carbon.identity.user.pre.update.password.action/src/main/java/org/wso2/carbon/identity/user/pre/update/password/action/api/model/Credential.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.user.pre.update.password.action.service.model;
+package org.wso2.carbon.identity.user.pre.update.password.action.api.model;
 
 /**
  * Represents an unencrypted credential.
