+++ components/user-mgt/org.wso2.carbon.identity.user.pre.update.password.action/src/main/java/org/wso2/carbon/identity/user/pre/update/password/action/api/model/PasswordSharing.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.user.pre.update.password.action.service.model;
+package org.wso2.carbon.identity.user.pre.update.password.action.api.model;
 
 import org.wso2.carbon.identity.certificate.management.model.Certificate;
