+++ components/user-mgt/org.wso2.carbon.identity.user.pre.update.password.action/src/main/java/org/wso2/carbon/identity/user/pre/update/password/action/internal/component/PreUpdatePasswordActionServiceComponentHolder.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.user.pre.update.password.action.internal;
+package org.wso2.carbon.identity.user.pre.update.password.action.internal.component;
 
 import org.wso2.carbon.identity.action.execution.ActionExecutorService;
 import org.wso2.carbon.identity.certificate.management.service.CertificateManagementService;
