+++ components/user-mgt/org.wso2.carbon.identity.user.pre.update.password.action/src/main/java/org/wso2/carbon/identity/user/pre/update/password/action/internal/constant/PreUpdatePasswordActionConstants.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.user.pre.update.password.action.core.constant;
+package org.wso2.carbon.identity.user.pre.update.password.action.internal.constant;
 
 /**
  * This class holds the constants used in Pre Update Password extension.
